// src/types/contentTypes.d.ts
export interface ApiPagePage extends Schema.CollectionType {
  collectionName: 'pages';
  info: {
    singularName: 'page';
    pluralName: 'pages';
    displayName: 'Forms';
    description: '';
  };
  options: {
    draftAndPublish: true;
  };
  attributes: {
    title: Attribute.String & Attribute.Required & Attribute.Unique;
    publish_at: Attribute.DateTime;
    slug: Attribute.UID & Attribute.Required;
    form: Attribute.DynamicZone<
      ['form.checkbox', 'form.select', 'form.textfield', 'form.file-uploader']
    > &
      Attribute.Required &
      Attribute.SetMinMax<{
        min: 1;
      }>;
    link: Attribute.String &
      Attribute.CustomField<'plugin::generated-page-link.page-link'>;
    description: Attribute.RichText & Attribute.Required;
    createdAt: Attribute.DateTime;
    updatedAt: Attribute.DateTime;
    publishedAt: Attribute.DateTime;
    createdBy: Attribute.Relation<'api::page.page', 'oneToOne', 'admin::user'> &
      Attribute.Private;
    updatedBy: Attribute.Relation<'api::page.page', 'oneToOne', 'admin::user'> &
      Attribute.Private;
  };
}